glTFとVRMのスキーマ定義に合わせて修正。

+++ src/Sansa.Model/AvatarTF/AvatarTF.Offset.cs
@@ -13,17 +13,17 @@
             /// <summary>
             /// TODO: 説明
             /// </summary>
-            public double? x { get; set; } = 0;
+            public float? x { get; set; } = 0;
 
             /// <summary>
             /// TODO: 説明
             /// </summary>
-            public double? y { get; set; } = 0;
+            public float? y { get; set; } = 0;
 
             /// <summary>
             /// TODO: 説明
             /// </summary>
-            public double? z { get; set; } = 0;
+            public float? z { get; set; } = 0;
 
 #pragma warning restore IDE1006 // 命名スタイル
         }
